requirements.txt
alibi==0.7.0
catboost==1.0.6
emoji==1.7.0
gensim==4.2.0
imblearn==0.0
joblib==1.1.0
Keras==2.9.0
keras_self_attention==0.51.0
lime==0.2.0.1
matplotlib==3.5.1
nltk==3.7
numpy==1.22.1
pandas==1.4.2
plotly==5.8.1
tweet-preprocessor==0.6.0
researchpy==0.3.2
responsible_ai_datacleaner==0.0.9
scikit_learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
shap==0.40.0
spacy==3.3.1
tensorflow==2.9.1
torch==1.11.0
tqdm==4.62.3
transformers==4.18.0
# tune_sklearn==0.4.3
xgboost==1.6.1
